Remove all direct access to maven, everything is accessed through meta.fabricmc.net now.
Still needs some tweaks to the meta api to provide the correct main class to the server installer
