+++ src/main/java/net/fabricmc/installer/util/LauncherMeta.java
@@ -18,12 +18,8 @@ package net.fabricmc.installer.util;
 
 import mjson.Json;
 
-import java.io.BufferedReader;
 import java.io.IOException;
-import java.io.InputStreamReader;
 import java.net.URL;
-import java.net.URLConnection;
-import java.nio.charset.StandardCharsets;
 import java.util.List;
 import java.util.stream.Collectors;
