+++ src/main/java/net/fabricmc/installer/util/Reference.java
@@ -17,10 +17,8 @@
 package net.fabricmc.installer.util;
 
 public class Reference {
-	public static final String PACKAGE = "net/fabricmc";
 	public static final String LOADER_NAME = "fabric-loader";
-	public static final String MAPPINGS_NAME = "intermediary";
 
-	public static String mavenServerUrl = "https://maven.fabricmc.net/";
+	public static String metaServerUrl = "https://meta.fabricmc.net/";
 	public static String fabricApiUrl = "https://www.curseforge.com/minecraft/mc-mods/fabric-api/";
 }
